+++ packages/merkleized-metadata/src/MerkleizedMetatada.ts
@@ -1,7 +1,8 @@
 import { $Metadata, Metadata, PortableRegistry, RuntimeVersion } from '@dedot/codecs';
-import { assert, blake3AsHex, HexString, stringCamelCase } from '@dedot/utils';
-import { $MetadataDigest } from './codecs';
-import { createMetadataDigest } from './digest.js';
+import { assert, blake3AsHex, blake3AsU8a, HexString, stringCamelCase } from '@dedot/utils';
+import { $ExtrinsicMetadata, $MetadataDigest, $TypeInfo, MetadataDigest } from './codecs';
+import { buildMerkleTree } from './merkle';
+import { transformMetadata } from './transform';
 import { ChainInfo, ChainInfoOptional, MetadataProof } from './types.js';
 
 /**
@@ -43,7 +44,33 @@ export class MerkleizedMetatada {
    * @returns The metadata digest
    */
   digest(): HexString {
-    return blake3AsHex($MetadataDigest.encode(createMetadataDigest(this.#metadata, this.#chainInfo)));
+    // Transform metadata to RFC format
+    const { typeInfo, extrinsicMetadata } = transformMetadata(this.#metadata);
+
+    // Encode type information
+    const encodedTypes = typeInfo.map((info) => $TypeInfo.encode(info));
+
+    // Build merkle tree from encoded type information
+    const typeTree = buildMerkleTree(encodedTypes);
+
+    // Hash extrinsic metadata
+    const encodedExtrinsicMetadata = $ExtrinsicMetadata.encode(extrinsicMetadata);
+    const extrinsicMetadataHash = blake3AsU8a(encodedExtrinsicMetadata);
+
+    const digest: MetadataDigest = {
+      type: 'V1',
+      value: {
+        typeInformationTreeRoot: typeTree.hash,
+        extrinsicMetadataHash,
+        specVersion: this.#chainInfo.specVersion,
+        specName: this.#chainInfo.specName,
+        base58Prefix: this.#chainInfo.ss58Prefix,
+        decimals: this.#chainInfo.decimals,
+        tokenSymbol: this.#chainInfo.tokenSymbol,
+      },
+    };
+
+    return blake3AsHex($MetadataDigest.encode(digest));
   }
 
   /**
