+++ packages/merkleized-metadata/src/codecs.ts
@@ -33,13 +33,8 @@ export type TypeRef = $.Input<typeof $TypeRef>;
  * Field in a composite or enumeration type
  */
 export const $Field = $.Struct({
-  /** Optional field name */
   name: $.Option($.str),
-
-  /** Type reference */
   ty: $TypeRef,
-
-  /** Optional type name */
   typeName: $.Option($.str),
 });
 
@@ -49,10 +44,7 @@ export type Field = $.Input<typeof $Field>;
  * Array type definition
  */
 export const $TypeDefArray = $.Struct({
-  /** Array length */
   len: $.u32,
-
-  /** Type parameter */
   typeParam: $TypeRef,
 });
 
@@ -62,10 +54,7 @@ export type TypeDefArray = $.Input<typeof $TypeDefArray>;
  * Bit sequence type definition
  */
 export const $TypeDefBitSequence = $.Struct({
-  /** Number of bytes */
   numBytes: $.u32,
-
-  /** Whether least significant bit is first */
   leastSignificantBitFirst: $.bool,
 });
 
@@ -75,13 +64,8 @@ export type TypeDefBitSequence = $.Input<typeof $TypeDefBitSequence>;
  * Enumeration variant
  */
 export const $EnumerationVariant = $.Struct({
-  /** Variant name */
   name: $.str,
-
-  /** Fields in the variant */
   fields: $.Vec($Field),
-
-  /** Variant index */
   index: $.compactU32,
 });
 
@@ -105,13 +89,8 @@ export type TypeDef = $.Input<typeof $TypeDef>;
  * Type information as defined in RFC-0078
  */
 export const $TypeInfo = $.Struct({
-  /** Type path */
   path: $.Vec($.str),
-
-  /** Type definition */
   typeDef: $TypeDef,
-
-  /** Type ID */
   typeId: $.compactU32,
 });
 
@@ -121,13 +100,8 @@ export type TypeInfo = $.Input<typeof $TypeInfo>;
  * Signed extension metadata as defined in RFC-0078
  */
 export const $SignedExtensionMetadata = $.Struct({
-  /** Identifier */
   identifier: $.str,
-
-  /** Type included in extrinsic */
   includedInExtrinsic: $TypeRef,
-
-  /** Type included in signed data */
   includedInSignedData: $TypeRef,
 });
 
@@ -137,19 +111,10 @@ export type SignedExtensionMetadata = $.Input<typeof $SignedExtensionMetadata>;
  * Extrinsic metadata as defined in RFC-0078
  */
 export const $ExtrinsicMetadata = $.Struct({
-  /** Extrinsic version */
   version: $.u32,
-
-  /** Address type */
   addressTy: $TypeRef,
-
-  /** Call type */
   callTy: $TypeRef,
-
-  /** Signature type */
   signatureTy: $TypeRef,
-
-  /** Signed extensions */
   signedExtensions: $.Vec($SignedExtensionMetadata),
 });
 
@@ -161,25 +126,12 @@ export type ExtrinsicMetadata = $.Input<typeof $ExtrinsicMetadata>;
 export const $MetadataDigest = $.Enum({
   V0: notSupportedCodec('Not supported!'),
   V1: $.Struct({
-    /** Root hash of the type information tree */
     typeInformationTreeRoot: $Bytes,
-
-    /** Hash of the extrinsic metadata */
     extrinsicMetadataHash: $Bytes,
-
-    /** Runtime spec version */
     specVersion: $.u32,
-
-    /** Runtime spec name */
     specName: $.str,
-
-    /** SS58 address format prefix */
     base58Prefix: $.u32,
-
-    /** Token decimal places */
     decimals: $.u32,
-
-    /** Token symbol */
     tokenSymbol: $.str,
   }),
 });
